Fix missing chr offset in mean and std load
Means and stds are loaded from text files with one float per line, one
line per marker. Previously, the start and end line indices for the read
operation were not offset for blocks on different chromosomes, in
effect, for all chromosomes except Chr1, wrong means and stds were
loaded. This is because in the block specifications, the marker indices
are not equivalent to the line index in the bim file, but indicate the
marker position on the chromosome.