rm checkEdges from rule 4
this is time critical; this function checks, for any edge (a, b) on a
discriminating path, if the edge can be deleted by conditioning on any
subset of the union of the sepset(X, Y) sepset(Y, X), where X and Y are the
endpoints of the discriminating path. With this modification we assume
that the skeleton and the sepsets that we obtain from the first stage
are sufficient and correct.